==> homework-5 - databases/2. BBC-final_project-data_acquisition.ipynb <==
import pandas as pd
import numpy as np
from urllib.request import urlopen
from bs4 import BeautifulSoup
import re

# Scrap the BBC page

raw_html = urlopen("http://www.bbc.com/culture/story/20160819-the-21st-centurys-100-greatest-films-who-voted").read()
soup_doc = BeautifulSoup(raw_html, "html.parser")

top_ten_list = []

paragraphs = (soup_doc
              .find('div', attrs={'class':'body-content'})
              .find_all('p', text=re.compile(r'^.*–.*\(.*\)$')))

for p in paragraphs:
    films = []
    for film in p.find_next_sibling():
        if film.name == None:
            films.append(re.search(r'^\d\d?\. (.*)', film).group(1))
    top_ten_list.append({
        'author': p.get_text(),
        'list': films
    })
    
top_ten_list[0]

# Create a flatten dataframe of all information available

df_critics.head()

reviews = []

for top_ten in top_ten_list:
    for index, review in enumerate(top_ten['list']):
        reviews.append({
            'critic': top_ten['author'],
            'film': review,
            'rank': index
        })

df_reviews = pd.DataFrame(reviews)

df_reviews.head()

# Correct all the duplications

df_reviews['film'].unique()

df_reviews.loc[df_reviews['film'] == 'Man on Wire (James Marsch, 2008)', 'film'] = 'Man on Wire (James Marsh, 2008)'
df_reviews.loc[df_reviews['film'] == 'Morvern Callar (Lynne Ramsay, 2012)', 'film'] = 'Morvern Callar (Lynne Ramsay, 2002)'
df_reviews.loc[df_reviews['film'] == 'Moolaadé (Ousmane, Sembèène 2004)', 'film'] = 'Moolaadé (Ousmane Sembène, 2004)'
df_reviews.loc[df_reviews['film'] == 'Fantastic Mr Fox (Wes Anderson, 2012)', 'film'] = 'Fantastic Mr Fox (Wes Anderson, 2009)'
df_reviews.loc[df_reviews['film'] == 'Finding Nemo (Andrew Stanton and Lee Unkrich, 2003)', 'film'] = 'Finding Nemo (Andrew Stanton, 2003)'

len(df_reviews['film'].unique())

# Create the PSQL tables

## Create the countries table

ISO_countries = pd.read_csv('iso_3166_2_countries.csv')[['Common Name', 'ISO 3166-1 2 Letter Code']]
ISO_countries.columns = ['common_name', 'code']
ISO_countries.index.name = 'id'
ISO_countries.to_csv('countries.csv')

## Create the reviews table

### Extract the films

films = []

def set_films(row):
    if row['film'] not in films:
        films.append(row['film'])
    return films.index(row['film'])
    
df_reviews['film'] = df_reviews.apply(set_films, axis=1)
df_reviews.rename(columns={'film':'film_id'}, inplace=True)

### Extract the critics

critics = []

def set_critics(row):
    if row['critic'] not in critics:
        critics.append(row['critic'])
    return critics.index(row['critic'])
    
df_reviews['critic'] = df_reviews.apply(set_critics, axis=1)
df_reviews.rename(columns={'critic':'critic_id'}, inplace=True)

df_reviews.head()

### Save the reviews table 

df_reviews.index.name = 'id'
df_reviews.to_csv('reviews.csv')

# Control if everything is ok with the csv exported from pandas
df_reviews = pd.read_csv('reviews.csv', index_col='id')
df_reviews.head()

# Control if everything is ok with the csv exported from psql
df_reviews = pd.read_csv('reviews-psql.csv', index_col='id')
df_reviews.head()

# Basic analytics
(df_reviews[df_reviews['rank'] <= 2]
 .groupby('film_id')['rank']
 .value_counts()
 .groupby(level=1)
 .nlargest(3))

## Create the Critics table

for index, critic in enumerate(critics):
    first_name = re.search(r'((.*)( .*)?) (.*) \–', critic).group(1)
    last_name = re.search(r'((.*)( .*)?) (.*) \–', critic).group(4)
    nationality = re.search(r'\((.*)\)', critic).group(1)
    company = re.search(r' – (.*) \(', critic).group(1)
    critics[index] = {
        'first_name': first_name, 
        'last_name': last_name,
        'nationality': nationality,
        'company': company
    }
    
df_critics = pd.DataFrame(critics)

df_critics.head()

# Merge the nationalities with the codes
# left: use only keys from left frame, similar to a SQL left outer join; preserve key order
df_critics = df_critics.merge(ISO_countries, left_on='nationality', right_on='common_name', how='left')

# create a list of the countries that have not been found
list(df_critics[df_critics['code'].isnull()]['nationality'].unique())

# Manually replace the missing countries unrecognized by the merge
countries = {
    'US':'US', 
    'UK':'GB', 
    'South Korea':'KR', 
    'UAE':'AE', 
    'Namibia':'NA', 
    'Taiwan':'TW', 
    'China':'CN'
}
# Copy the country name to the country code for the unrecognized rows
df_critics.loc[df_critics['code'].isnull(), 'code'] = df_critics[df_critics['code'].isnull()]['nationality']
# Replace the non-standard codes
df_critics['code'].replace(countries, inplace=True)

# Get rid of the non-standard country columns
df_critics = df_critics.drop('nationality', 1)
df_critics = df_critics.drop('common_name', 1)

# Rename the country_code column header
df_critics.rename(columns={'code':'nationality_code'}, inplace=True)

df_critics.head()

df_critics.index.name = 'id'
df_critics.to_csv('critics.csv')

## Create the films table

directors_list = []

for index, film in enumerate(films):
    title = re.search(r'(.*) \(', film).group(1)
    director = re.search(r'\(((.(?!\) \())+), [0-9]{4}\)$', film).group(1)
    release_year = int(re.search(r'\((.*), ([0-9]{4})', film).group(2))
    directors_list.append(director)
    films[index] = {
        'title': title, 
        'release_year': release_year
    }
    
df_films = pd.DataFrame(films)
df_films.index.name = 'id'
df_films.to_csv('films.csv')

df_films.head()

## Create the directions and directors table

directions = []
directors = []

for index, d in enumerate(directors_list):
    for director in re.findall(r"[^,]+", d):
        for dd in director.split(' and '):
            if dd not in directors:
                directors.append(dd.lstrip())
            directions.append({
                'director_id': directors.index(dd.lstrip()),
                'film_id': index
            })
            
df_directions = pd.DataFrame(directions)
df_directions.index.name = 'id'
df_directions.to_csv('directions.csv')

df_directions.head()

for index, director in enumerate(directors):
    if re.search(r'((.*)( .*)?) (.*)$', director) == None:
        first_name = director
        last_name = np.nan
    else:
        first_name = re.search(r'((.*)( .*)?) (.*)$', director).group(1)
        last_name = re.search(r'((.*)( .*)?) (.*)$', director).group(4)
        
    directors[index] = {
        'first_name': first_name, 
        'last_name': last_name
    }
    
df_directors = pd.DataFrame(directors)

# Check for errors
df_directors[df_directors['last_name'].isnull()]

# Manually correct the missing last names
df_directors.loc[(df_directors['first_name'] == 'Joel') & df_directors['last_name'].isnull(), 'last_name'] = 'Cohen'
df_directors.loc[(df_directors['first_name'] == 'Jean-Pierre') & df_directors['last_name'].isnull(), 'last_name'] = 'Dardenne'
df_directors.loc[(df_directors['first_name'] == 'Josh') & df_directors['last_name'].isnull(), 'last_name'] = 'Safdie'
df_directors.loc[(df_directors['first_name'] == 'Anthony') & df_directors['last_name'].isnull(), 'last_name'] = 'Russo'

# Check again
df_directors[df_directors['last_name'].isnull()]

# Save the directors table
df_directors = pd.DataFrame(directors)
df_directors.index.name = 'id'
df_directors.to_csv('directors.csv')

# Manually correct the comma misplacements in the csv

df_directors = pd.read_csv('directors.csv', index_col='id')

df_directors.head()

# All the tables

print('\n--------------------------')
print('Critics')
print('--------------------------')
print(df_critics.dtypes)
print('\n--------------------------')
print('Films')
print('--------------------------')
print(df_films.dtypes)
print('\n--------------------------')
print('Directors')
print('--------------------------')
print(df_directors.dtypes)
print('\n--------------------------')
print('Reviews')
print('--------------------------')
print(df_reviews.dtypes)
print('\n--------------------------')
print('Directions')
print('--------------------------')
print(df_directions.dtypes)

# All the films directed by 3 or more people
# PSQL
# select film_id, count(film_id) from directions group by film_id having count(film_id) > 2;
df_directions.groupby('film_id').filter(lambda x: len(x) > 2)['film_id'].value_counts()
