==> homework-5 - databases/5. BBC-final_project-refining_selenium.ipynb <==
import pandas as pd
import numpy as np
import re

df_critics = pd.read_csv('critics_final.csv', index_col='id')
df_films = pd.read_csv('films_selenium_raw.csv', index_col='id')
df_reviews = pd.read_csv('reviews_final.csv', index_col='id')
df_directors = pd.read_csv('directors_selenium_raw.csv', index_col='id')
df_directions = pd.read_csv('directions_final.csv', index_col='id')
df_genres = pd.read_csv('genres_final.csv', index_col='id')
df_countries = pd.read_csv('countries-psql.csv', index_col='id', keep_default_na=False)
df_continents = pd.read_csv('continents-psql.csv', index_col='id', keep_default_na=False)

df_genres[df_genres['film_id'] == 0]

df_merged = (df_reviews
 .merge(df_critics, left_on='critic_id', right_index=True)
 .merge(df_films, left_on='film_id', right_index=True)
 .merge(df_directions, left_on='film_id', right_on='film_id')
 .merge(df_directors, left_on='director_id', right_index=True))

df_merged[df_merged['full_name'] == 'Alyssa Rosenberg'].sort_values('rank')

df_merged[df_merged['film_id'] == 393]

df_genres[df_genres['genre'] == 'Short'].merge(df_films, left_on='film_id', right_index=True)

df_directors = pd.read_csv('directors_selenium_raw.csv')

df_directors[75:85]

df_merged = df_directors.merge(df_countries, left_on='nationality', right_on='name', how='left')

df_merged[df_merged['code'].isnull()]['nationality'].unique()

countries = {
    'UK':'GB', 
    'USA':'US', 
    'Palestine':'PS', 
    'Taiwan':'TW', 
    'China':'CN',
    'Yugoslavia':'MK',
    'South Korea':'KR'
}

df_merged.loc[df_merged['code'].isnull(), 'code'] = df_merged['nationality']

df_merged['code'].replace(countries, inplace=True)

df_merged[75:85]

df_merged[['id', 'name_x', 'year_of_birth', 'code']].to_csv('directors_final.csv', index=False)

df_films = pd.read_csv('films_selenium_raw.csv')

df_films.head()

df_films['duration'] = df_films['duration'].str.extract('PT(\d*)M')

df_films.head()

df_films.to_csv('films_final.csv', index=False)
